fix: widget close functionality and major refactoring
- Fixed creation widget close button not working (added widget_closed event handler)
- Fixed focus management after widget actions (returns focus to chat input)
- Moved PreviewWidget close option to context menu for cleaner UI
- Refactored PreviewWidget and LoginWidget into folder structures with smaller components
- Cleaned up unused widget folders and abandoned _base architecture
- Fixed all linting errors and warnings (nullish coalescing, semantic colors)
- Removed duplicate/unused widget implementations
