+++ src/app/auth/verify-success/page.tsx
@@ -27,21 +27,21 @@ export default function VerifySuccessPage() {
   return (
     <div className="min-h-screen flex items-center justify-center bg-gradient-to-br from-primary-50 to-secondary-50 dark:from-primary-950 dark:to-secondary-950">
       <div className="max-w-md w-full mx-auto p-6">
-        <div className="bg-white dark:bg-gray-900 rounded-lg shadow-xl p-8 text-center">
+        <div className="bg-white dark:bg-neutral-900 rounded-lg shadow-xl p-8 text-center">
           <div className="flex justify-center mb-4">
-            <CheckCircle className="h-16 w-16 text-green-500" />
+            <CheckCircle className="h-16 w-16 text-success" />
           </div>
           
-          <h1 className="text-2xl font-bold text-gray-900 dark:text-gray-100 mb-2">
+          <h1 className="text-2xl font-bold text-neutral-900 dark:text-neutral-100 mb-2">
             Email Verified Successfully!
           </h1>
           
-          <p className="text-gray-600 dark:text-gray-400 mb-6">
+          <p className="text-neutral-600 dark:text-neutral-400 mb-6">
             Your email has been verified. You can now log in to your account.
           </p>
           
           <div className="space-y-4">
-            <p className="text-sm text-gray-500 dark:text-gray-500">
+            <p className="text-sm text-neutral-500 dark:text-neutral-500">
               Redirecting to Hexframe in {countdown} seconds...
             </p>
