+++ src/app/map/Chat/Widgets/PreviewWidget/index.tsx
